fix: prevent vectorscope Uint16 overflow and fix conditional hook calls
Change vectorscope density buffer from Uint16Array to Uint32Array to
handle paused sample sizes (384x216 = 82,944 pixels) that can overflow
a single bin. Move visibleTrackIds and activeMaskInfos useMemo hooks
before the early return in CompositionContent so hooks run
unconditionally.

+++ src/shared/components/color-scopes-view.tsx
@@ -281,7 +281,7 @@ function drawVectorscope(
   const { kr, kb } = getMatrixCoefficients(matrix);
   const kg = 1 - kr - kb;
   const { data, width, height } = imageData;
-  const density = new Uint16Array(scopeSize * scopeSize);
+  const density = new Uint32Array(scopeSize * scopeSize);
   let maxDensity = 0;
 
   const center = Math.floor(scopeSize / 2);
